tech: annotations refactor, add tests for regions processing
